requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.2.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.7
cryptography==37.0.1
decorator==5.1.1
Deprecated==1.2.13
facebook-wda==1.4.6
Faker==9.7.1
hrpc==1.0.9
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.1
jsonpath==0.82
MarkupSafe==2.1.1
mss==6.1.0
numpy
opencv-contrib-python==4.5.5.64
packaging==21.2
Pillow==10.3.0
pluggy==1.0.0
py==1.10.0
pycparser==2.21
PyHamcrest==2.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-rerunfailures==10.2
python-dateutil==2.8.2
python-xlib==0.31
pywinauto==0.6.3
PyYAML==6.0.1
requests==2.26.0
retry==0.9.2
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
websocket-client==0.48.0
wrapt==1.14.0
pytest-dependency
pytest-ordering